:art: Adding LostMaterial's List Feature
The Adding LostMaterial's List Feature allow users to list all, returned or not returned lost materials. The implementation was made using the laravel querys. It was made the choice between high performance or better readability of the code and was choosed the readability due to the tiny quantity of people that will indirectly use the application.